app/digital_advisor/pyproject.toml
[project]
name = "digital-advisor"
version = "0.1.0"
description = ""
authors = [
    {name = "Razvan Nica",email = "242715@buas.nl"}
]
readme = "README.md"
requires-python = ">=3.11 , <3.12"
dependencies = [
    "numpy (==1.26.0)",
    "pandas (==2.0.3)",
    "scikit-learn (==1.6.1)",
    "joblib (==1.5.1)",
    "tensorflow-io-gcs-filesystem (>=0.28.0,<0.30.0)",
    "tensorflow (==2.18.0)",
    "marshmallow (>=4.0.0,<5.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]

[tool.poetry]
packages = [{include = "digital_advisor", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
